moved Client type to types.ts
