+++ client/src/components/clientPage/ViewPage.tsx
@@ -20,49 +20,7 @@ import { useBackendContext } from "../../contexts/hooks/useBackendContext";
 import Comments from "./comments";
 import ChildrenCards from "./childrenCards";
 import Forms from "./forms"
-import { InitialInterview, ExitSurvey, FormItems, SuccessStory } from "./types";
-interface Client {
-  age: number;
-  attendingSchoolUponEntry: boolean;
-  attendingSchoolUponExit: boolean;
-  bedNights: number;
-  bedNightsChildren: number;
-  chronicallyHomeless: boolean;
-  cityOfLastPermanentResidence: string;
-  createdBy: number;
-  dateOfBirth: string;
-  destinationcity: string;
-  disabledChildren: boolean;
-  email: string;
-  emergencyContactName: string;
-  emergencyContactPhoneNumber: string;
-  employmentGained: boolean;
-  entranceDate: string;
-  estimatedExitdate: string;
-  ethnicity: string;
-  exitDate: string;
-  firstName: string;
-  grant: string;
-  homelessnessLength: number;
-  id: number;
-  lastName: string;
-  medical: boolean;
-  phoneNumber: string;
-  pregnantUponEntry: boolean;
-  priorLiving: string;
-  priorLivingCity: string;
-  race: string;
-  reasonForLeaving: string;
-  reunified: boolean;
-  savingsAmount: string;
-  shelterInLastFiveYears: boolean;
-  specificDestination: string;
-  specificReasonForLeaving: string;
-  status: string;
-  successfulCompletion: boolean;
-  unitId: number;
-  comments: string;
-}
+import { Client, FormItems } from "./types";
 
 const emptyClient: Client = {
   age: 0,
